refactoring delete into its own file
